chore: 🔧 줄간격 수정 및 FiterUseCase 분리(#148)
