+++ ios/IssueTracker/IssueTracker/NewIssue/UseCase/NewIssueFilterUseCase.swift
@@ -98,4 +98,5 @@ extension NewIssueFilterUseCase {
             return children
         }
     }
+    
 }
