+++ ios/IssueTracker/IssueTracker/NewIssue/UseCase/PostNewIssueUseCase.swift
@@ -36,4 +36,5 @@ final class DefaultPostNewIssueUseCase: PostNewIssueUseCase {
             .store(in: &cancelBag)
 
     }
+    
 }
