+++ ios/IssueTracker/IssueTracker/NewIssue/UseCase/UploadImageUseCase.swift
@@ -34,4 +34,5 @@ final class DefaultUploadImageUseCase: UploadImageUseCase {
             }.store(in: &cancelBag)
 
     }
+    
 }
